backend/requirements.txt
fastapi
uvicorn
sqlalchemy
httpx
slowapi
passlib[bcrypt]
bcrypt==4.0.1
email-validator
python-jose[cryptography]
python-dotenv